add repeat email field and refactor
