+++ assets/lang/fr-FR.json
@@ -32,7 +32,10 @@
             "button-submit": "Changer d'adresse e-mail maintenant",
             "label-description": "Veuillez saisir votre nouvelle adresse e-mail.",
             "label-title": "Modifier l'adresse e-mail",
-            "hint-label": "Nouvelle adresse e-mail"
+            "hint-label": "Nouvelle adresse e-mail",
+            "hint-label-repeat": "Répéter la nouvelle adresse e-mail",
+            "email-repeat-mismatch-error": "Les adresses e-mail ne correspondent pas",
+            "email-not-valid": "Veuillez saisir une adresse e-mail valide"
         }
     },
     "alertdialog": {
