--- lib/core/authentication/presentation/widgets/change_email_action_lip/change_email_action_lip_scaffold.dart
@@ -1,91 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:gradient_widgets/gradient_widgets.dart';
-import 'package:linum/core/design/layout/utils/layout_helpers.dart';
-
-class ChangeEmailActionLipScaffold extends StatelessWidget {
-  final String label;
-  final Widget inputField;
-  final VoidCallback callback;
-  final String buttonLabel;
-  final TextEditingController controller;
-  const ChangeEmailActionLipScaffold({
-    super.key,
-    required this.label,
-    required this.inputField,
-    required this.controller,
-    required this.callback,
-    required this.buttonLabel,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(
-            vertical: 12.0,
-            horizontal: 24.0,
-          ),
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0),
-            child: Text(
-              label,
-              style: Theme.of(context).textTheme.bodyLarge,
-              textAlign: TextAlign.center,
-            ),
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            children: [
-              Container(
-                padding: const EdgeInsets.all(5.0),
-                decoration: BoxDecoration(
-                  color: Theme.of(context).colorScheme.onSecondary,
-                  borderRadius: BorderRadius.circular(10),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Theme.of(context).colorScheme.onBackground,
-                      blurRadius: 20.0,
-                      offset: const Offset(0, 10),
-                    ),
-                  ],
-                ),
-                child: Column(
-                  children: [
-                    inputField,
-                  ],
-                ),
-              ),
-              SizedBox(
-                height: context.proportionateScreenHeight(32),
-              ),
-              GradientButton(
-                increaseHeightBy:
-                context.proportionateScreenHeight(16),
-                // Logged Out onPressed
-                callback: callback,
-                gradient: LinearGradient(
-                  colors: [
-                    Theme.of(context).colorScheme.primary,
-                    Theme.of(context).colorScheme.surface,
-                  ],
-                ),
-                elevation: 0,
-                increaseWidthBy: double.infinity,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: Text(
-                  buttonLabel,
-                  style: Theme.of(context).textTheme.labelLarge,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
